+++ webpack.common.js
@@ -1,28 +1,13 @@
 const path = require("path");
 
 const HtmlWebpackPlugin = require("html-webpack-plugin");
-const CopyPlugin = require("copy-webpack-plugin");
 
 module.exports = {
-  mode: "development",
   entry: path.join(__dirname, "src", "index.tsx"),
-  devtool: "inline-source-map",
   output: {
-    path: path.resolve(__dirname, "dist"),
+    path: path.join(__dirname, "dist"),
     filename: "[name].[contenthash].js",
     chunkFilename: "[name].[contenthash].js",
-  },
-  optimization: {
-    runtimeChunk: "single",
-    splitChunks: {
-      chunks: "all",
-    },
-  },
-  devServer: {
-    static: path.join(__dirname, "public"),
-    port: 3000,
-    hot: true,
-    historyApiFallback: true,
   },
   module: {
     rules: [
@@ -47,16 +32,6 @@ module.exports = {
   plugins: [
     new HtmlWebpackPlugin({
       template: path.join(__dirname, "public", "index.html"),
-    }),
-    new CopyPlugin({
-      patterns: [
-        {
-          from: "public",
-          globOptions: {
-            ignore: ["**/index.html"],
-          },
-        },
-      ],
     }),
   ],
 };